pyproject.toml
[project]
name = "framework-python"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
  "alembic>=1.18.4",
  "bcrypt<4.0.0",
  "email-validator>=2.3.0",
  "fastapi[standard]>=0.129.0",
  "jinja2>=3.1.6",
  "mjml>=0.12.0",
  "passlib>=1.7.4",
  "psycopg[binary]>=3.3.3",
  "pydantic-settings>=2.12.0",
  "pyjwt>=2.11.0",
  "sqlalchemy>=2.0.46",
  "sqlmodel>=0.0.34",
]

[dependency-groups]
dev = [
  "pre-commit>=4.5.1",
  "ruff>=0.15.2",
]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
]
ignore = [
  "E501",  # line too long, handled by black
  "B008",  # do not perform function calls in argument defaults
  "W191",  # indentation contains tabs
  "B904",  # Allow raising exceptions without from e, for HTTPException
  "E711",  # SQLAlchemy comparison to None
  "E712",  # SQLAlchemy comparison to True/False
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.per-file-ignores]
# Alembic revisions might have unused imports
"app/alembic/versions/*" = ["F401", "E501", "E402"]
